requirements.txt
django
Pillow==5.2.0
djangorestframework==3.9.1
